refactor: move tracklist models into own folder
